requirements.txt
helper==2.5.0
matplotlib
numpy==1.26.1
nvidia_ml_py==12.570.86
opencv_python==4.11.0.86
pandas==2.2.3
Pillow==12.1.0
PyYAML==6.0.3
qwen_vl_utils==0.0.14
Requests==2.32.5
scikit_learn
scipy
seaborn==0.13.2
setuptools==75.8.0
timm
tqdm==4.67.1
transformers==4.49.0
https://github.com/kyamagu/faiss-wheels/releases/download/v1.7.3/faiss_gpu-1.7.3-cp310-cp310-manylinux_2_17_x86_64.manylinux2014_x86_64.whl